+++ packages/router-plugin/src/core/router-code-splitter-plugin.ts
@@ -20,7 +20,7 @@ import {
   tsrShared,
   tsrSplit,
 } from './constants'
-import { debug, normalizePath } from './utils'
+import { debug, normalizePath, routeFactoryCallCodeFilter } from './utils'
 import { createRouterPluginContext } from './router-plugin-context'
 import type { CodeSplitGroupings, SplitRouteIdentNodes } from './constants'
 import type { GetRoutesByFileMapResultValue } from '@tanstack/router-generator'
@@ -242,11 +242,6 @@ export function createRouterCodeSplitterPlugin(
     return result
   }
 
-  const includedCode = [
-    'createFileRoute(',
-    'createRootRoute(',
-    'createRootRouteWithContext(',
-  ]
   return [
     {
       name: 'tanstack-router:code-splitter:compile-reference-file',
@@ -260,17 +255,14 @@ export function createRouterCodeSplitterPlugin(
             include: /\.(m|c)?(j|t)sx?$/,
           },
           code: {
-            include: includedCode,
+            include: routeFactoryCallCodeFilter,
           },
         },
         handler(code, id) {
           const normalizedId = normalizePath(id)
           const generatorFileInfo =
             routerPluginContext.routesByFile.get(normalizedId)
-          if (
-            generatorFileInfo &&
-            includedCode.some((included) => code.includes(included))
-          ) {
+          if (generatorFileInfo) {
             return handleCompilingReferenceFile(
               code,
               normalizedId,