+++ packages/solid-start/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @tanstack/solid-start
 
+## 1.168.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/solid-router@1.170.4
+  - @tanstack/solid-start-client@1.167.4
+  - @tanstack/solid-start-server@1.167.4
+  - @tanstack/start-client-core@1.169.4
+  - @tanstack/start-plugin-core@1.170.6
+  - @tanstack/start-server-core@1.168.4
+
+## 1.168.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/start-plugin-core@1.170.5
+
+## 1.168.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/solid-router@1.170.3
+  - @tanstack/solid-start-client@1.167.3
+  - @tanstack/solid-start-server@1.167.3
+  - @tanstack/start-client-core@1.169.3
+  - @tanstack/start-plugin-core@1.170.4
+  - @tanstack/start-server-core@1.168.3
+
+## 1.168.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/solid-router@1.170.2
+  - @tanstack/solid-start-client@1.167.2
+  - @tanstack/solid-start-server@1.167.2
+  - @tanstack/start-client-core@1.169.2
+  - @tanstack/start-plugin-core@1.170.3
+  - @tanstack/start-server-core@1.168.2
+
+## 1.168.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/start-plugin-core@1.170.2
+
 ## 1.168.1
 
 ### Patch Changes