+++ packages/start-static-server-functions/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @tanstack/start-static-server-functions
 
+## 1.167.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/start-client-core@1.169.4
+  - @tanstack/react-start@1.168.6
+  - @tanstack/solid-start@1.168.6
+
+## 1.167.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/start-client-core@1.169.3
+  - @tanstack/react-start@1.168.4
+  - @tanstack/solid-start@1.168.4
+
+## 1.167.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/start-client-core@1.169.2
+  - @tanstack/react-start@1.168.3
+  - @tanstack/solid-start@1.168.3
+
 ## 1.167.1
 
 ### Patch Changes